Use a specialization for using C++ promise.
This workaround is created due to incompatibilities with HPUX
