+++ src/shared_modules/rsync/CMakeLists.txt
@@ -48,6 +48,9 @@ include_directories(${SHARED_MODULES}/common/)
 
 if(CMAKE_SYSTEM_NAME STREQUAL "HP-UX")
   link_directories(${INSTALL_PREFIX}/lib)
+  add_definitions(-DPROMISE_TYPE=PromiseType::SLEEP)
+  else()
+    add_definitions(-DPROMISE_TYPE=PromiseType::NORMAL)
 endif(CMAKE_SYSTEM_NAME STREQUAL "HP-UX")
 
 link_directories(${SHARED_MODULES}/dbsync/build/lib/)
